// src/api/generated/models/shippingInvoice.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { ShippingInvoiceCreatedByUserId } from './shippingInvoiceCreatedByUserId';
import type { ShippingInvoiceStatus } from './shippingInvoiceStatus';
import type { ShippingInvoiceTrackingNumber } from './shippingInvoiceTrackingNumber';
import type { ShippingInvoiceCarrier } from './shippingInvoiceCarrier';
import type { ShippingInvoiceNotes } from './shippingInvoiceNotes';
import type { ShippingInvoiceIssuedAt } from './shippingInvoiceIssuedAt';
import type { ShippingInvoiceInTransitAt } from './shippingInvoiceInTransitAt';
import type { ShippingInvoiceDeliveredAt } from './shippingInvoiceDeliveredAt';
import type { ShippingInvoiceBuyerCompany } from './shippingInvoiceBuyerCompany';
import type { ShippingInvoiceSupplierCompany } from './shippingInvoiceSupplierCompany';
import type { ShippingInvoiceManagedByCompany } from './shippingInvoiceManagedByCompany';
import type { ShippingInvoiceSupplierInvoice } from './shippingInvoiceSupplierInvoice';
import type { ShippingLine } from './shippingLine';

export interface ShippingInvoice {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  supplierInvoiceId: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  buyerCompanyId: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  supplierCompanyId: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  managedByCompanyId: string;
  createdByUserId: ShippingInvoiceCreatedByUserId;
  status: ShippingInvoiceStatus;
  trackingNumber: ShippingInvoiceTrackingNumber;
  carrier: ShippingInvoiceCarrier;
  notes: ShippingInvoiceNotes;
  issuedAt: ShippingInvoiceIssuedAt;
  inTransitAt: ShippingInvoiceInTransitAt;
  deliveredAt: ShippingInvoiceDeliveredAt;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  updatedAt: string;
  buyerCompany: ShippingInvoiceBuyerCompany;
  supplierCompany: ShippingInvoiceSupplierCompany;
  managedByCompany: ShippingInvoiceManagedByCompany;
  supplierInvoice: ShippingInvoiceSupplierInvoice;
  lines: ShippingLine[];
}